requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
setuptools==47.1.0
Werkzeug==0.16.1
SQLAlchemy==1.3.20
Flask-SQLAlchemy==2.4.4
cryptography==3.2.1
PyMySQL==0.10.1
Flask-WTF==0.14.3
google-api-python-client==1.12.5
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
google-auth==1.23.0
google-auth-httplib2==0.0.4
pytz==2021.1